Refactor: categoryList
